Migrating to .NET Standard

+++ src/MassTransit.EventStoreIntegration/EventsData.cs
@@ -4,13 +4,13 @@ namespace MassTransit.EventStoreIntegration
 {
     public class EventsData
     {
-        public EventsData(IEnumerable<object> events, int lastVersion)
+        public EventsData(IEnumerable<object> events, long lastVersion)
         {
             Events = events;
             LastVersion = lastVersion;
         }
 
         public IEnumerable<object> Events { get; }
-        public int LastVersion { get; }
+        public long LastVersion { get; }
     }
 }
